Splitting Buffer Primitive classes/etc into separate files (#1463)
* Splitting classes into separate files

* Removing extra file

* Splitting ManagedBufferPool into partial classes

* Removing duplicate code left from split

* Splitting NativeBufferPool into partial classes

* Fixing typo

* Addressing PR comments.
