+++ src/System.Buffers.Primitives/System/Internal/ManagedBufferPool.cs
