+++ src/System.Buffers.Primitives/System/Runtime/ReadOnlyBufferDebuggerView.cs
@@ -6,24 +6,6 @@ using System.Diagnostics;
 
 namespace System.Runtime
 {
-    internal class BufferDebuggerView<T>
-    {
-        private ReadOnlyBuffer<T> _buffer;
-
-        public BufferDebuggerView(Buffer<T> buffer)
-        {
-            _buffer = buffer;
-        }
-
-        [DebuggerBrowsable(DebuggerBrowsableState.RootHidden)]
-        public T[] Items
-        {
-            get {
-                return _buffer.ToArray();
-            }
-        }
-    }
-
     internal class ReadOnlyBufferDebuggerView<T>
     {
         private ReadOnlyBuffer<T> _buffer;
